better reporting of GLSL compile messages
If there are warning messages, but no error messages, we now report
the warning messages.

We report the pathname of the GLSL source file that contains the warnings
or errors (in case more than one shader is provided).

Line numbers reported in warning or error messages are now more accurate.
After the automatically generated prolog, a '#line 1' directive
is now inserted.

Fixed warnings in "mandelbrot.frag".