Show all GLSL warnings
This PR contains 3 of the 4 changes from "better reporting of GLSL
compile messages", and will be easier to code review.
 1. If there are warning messages, but no error messages,
    we now report the warning messages.
 2. Line numbers reported in warning or error messages are now more accurate.
    After the automatically generated prolog, a '#line 1' directive
    is now inserted.
 3. Fixed warnings in "mandelbrot.frag".

What's missing is the code that prints the path of the shader file
in the case that there are compile messages for the shader.
Since this is the code that you were concerned about during review of PR #51.