3dview: a viewer for ray-marched 3D shapes
Suppose you create a 3D scene in a fragment shader, using Function
Representation and ray-marching instead of polygons, as seen on shadertoy.com.

For this use case, I have created a 3D browser that allows you to pan, zoom,
and rotate the 3D scene, using the mouse.

The fragment shader uses 3 uniform variables:
```
  uniform vec3 u_eye3d;
  uniform vec3 u_centre3d;
  uniform vec3 u_up3d;
```
These are the same quantities that are passed to the `lookAt` function in GLM
and other libraries. See the examples `menger.frag` and `raymarch.frag`.

The user interface uses the scroll wheel for zoom, left-button for rotate,
and right-button for pan.