Fix screenshot saving from wrong buffer.
For double-buffered setups, the screenshot feature does not work because
glReadPixels is reading from the back buffer, which has already been
swapped by the time the process exits. There are some cases it works
though, in particular, if onScreenshot is triggered by the drawDone
event because it’s called before the buffer swap is done.

This fixes the problem by ensuring that the front buffer is what always
gets read for screenshot saving.