modify command line argument logic to prevent segfaults
If a command line argument which requires a value (e.g. `-x <pixels>`) is called without the second argument, this can cause a segfault because the `i` counter for iterating through `argv` is not checked against `argc` in the parser.

This patch adds error handling to the command line parser to prevent segfaults.