CMakeLists.txt: Avoid hardcoding installation dirs
This allows the user to specify alternative directories when
`/usr/share` is not a suitable option.

Also, remove `sudo` invocations. It should be more preferable to let the
user do `sudo cmake --install <build-directory>` instead, if needed.
This is safer, and allows users without root privileges to build and
install to custom locations.

See:
- https://cmake.org/cmake/help/latest/module/GNUInstallDirs.html
- https://www.gnu.org/prep/standards/html_node/Directory-Variables.html

Signed-off-by: Ruoyu Zhong <zhongruoyu@outlook.com>